Functions and Aggregates
This exercise aims to give you practice in:

defining views based on simple SQL queries
defining SQL functions
defining PLpgSQL functions
adding new aggregate operators
This exercise will not explain how to do everything in fine detail. Part of the aim of the exercise is that you explore how to use the PostgreSQL system. A very important tool for this is the PostgreSQL Documentation. In particular, the following sections of the Documentation are relevant:

the SQL CREATE FUNCTION statement;
the SQL CREATE AGGREGATE statement;
the PLpgSQL language description.